+++ src/Farmer/Arm/ContainerInstance.fs
@@ -14,6 +14,11 @@ type ContainerGroupIpAddress =
         {| Protocol : TransmissionProtocol
            Port : uint16 |} Set }
 
+type ImageRegistryCredential =
+    { Server : string
+      Username : string
+      Password : SecureParameter }
+
 type EnvVarValue = EnvValue of string | EnvSecureValue of string
 
 type ContainerGroup =
@@ -30,6 +35,7 @@ type ContainerGroup =
         |} list
       OperatingSystem : OS
       RestartPolicy : RestartPolicy
+      ImageRegistryCredentials : ImageRegistryCredential list
       IpAddress : ContainerGroupIpAddress
       NetworkProfile : ResourceName option
       Volumes : Map<string, Volume>
@@ -51,6 +57,8 @@ type ContainerGroup =
                 ()
     ]
 
+    interface IParameters with
+        member this.SecureParameters = this.ImageRegistryCredentials |> List.map (fun c -> c.Password)
     interface IArmResource with
         member this.ResourceName = this.Name
         member this.JsonModel =
@@ -83,6 +91,11 @@ type ContainerGroup =
                             | AlwaysRestart -> "Always"
                             | NeverRestart -> "Never"
                             | RestartOnFailure -> "OnFailure"
+                          imageRegistryCredentials =
+                              this.ImageRegistryCredentials |> List.map (fun cred ->
+                                  {| server = cred.Server
+                                     username = cred.Username
+                                     password = cred.Password.AsArmRef.Eval() |})
                           ipAddress =
                             {| ``type`` =
                                 match this.IpAddress.Type with